Changed the method for our 'q-drop' approach
Instead of finding a local maximum and then looking for qualities 30% below
that (exon alignment qualities), we now look for neighbors who show a
steep quality dropoff.  This avoids issues where we might end up slowly
decreasing in exon quality in such a way that the difference between the local
maximum and a low-quality alignment is misleading.