Added a check for previously used exons
We'll no longer allow any BLAT hits that overlap with previously identified
exonic regions to be considered as candidates for ghost exons.  We were picking
up too many hits for sequences that were technically not solid alignments
(when only considering identical / non-identical characters) but were similar
enough (thinking along the lines of similarity scores) to be re-identified as
correctly mapped.

Very awkward to word this.  The short version is, "looks sorta bad but seems
right under scrutiny" applies to quite a few of our transitively-aligned exons.